requirements.txt
flask
mysql-connector-python


#para executar coloque "pip install -r requirements.txt" no terminal